backend/requirements.txt
Flask==3.0.3
flask-cors==6.0.1
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==2.1.2
rembg==2.0.67
onnxruntime==1.23.0
requests==2.32.3
Werkzeug==3.0.4
flask-pymongo==3.0.1
